Add Authorise feature to the API.

+++ AndroidWebService/Controllers/Reservations/LovedMotelsController.cs
@@ -6,6 +6,8 @@ using AndroidWebService.Models;
 using System.Collections.Generic;
 using System.Web.Http.Description;
 using System.Data.Entity.Infrastructure;
+using System.Net.Http.Headers;
+using System.Net.Http;
 
 namespace AndroidWebService.Controllers.Reservations
 {
@@ -18,9 +20,15 @@ namespace AndroidWebService.Controllers.Reservations
         [HttpGet]
         public async Task<List<PTYeuThich>> GetByUserName(string userName)
         {
-            List<PTYeuThich> lovedMotels = await db.PTYeuThich.ToListAsync();
+            CookieHeaderValue cookie = Request.Headers.GetCookies("cookie-header").FirstOrDefault();
+            if (cookie != null)
+            {
+                List<PTYeuThich> lovedMotels = await db.PTYeuThich.ToListAsync();
 
-            return lovedMotels.Where(k => k.TenDangNhap.Trim() == userName.Trim()).ToList();
+                return lovedMotels.Where(k => k.TenDangNhap.Trim() == userName.Trim()).ToList();
+            }
+                
+            return new List<PTYeuThich>();    
         }
         [ResponseType(typeof(PTYeuThich))]
         [HttpGet]
@@ -35,44 +43,56 @@ namespace AndroidWebService.Controllers.Reservations
         [HttpPost]
         public async Task<IHttpActionResult> Post(PTYeuThich lovedMotel)
         {
-            if (!ModelState.IsValid)
+            CookieHeaderValue cookie = Request.Headers.GetCookies("cookie-header").FirstOrDefault();
+            if (cookie != null)
             {
-                return BadRequest(ModelState);
+                if (!ModelState.IsValid)
+                {
+                    return BadRequest(ModelState);
+                }
+
+                db.PTYeuThich.Add(lovedMotel);
+
+                try
+                {
+                    await db.SaveChangesAsync();
+                }
+                catch (DbUpdateException)
+                {
+                    throw;
+                }
+
+                return CreatedAtRoute("DefaultApi", new { id = lovedMotel.MaPT }, lovedMotel);
             }
 
-            db.PTYeuThich.Add(lovedMotel);
-
-            try
-            {
-                await db.SaveChangesAsync();
-            }
-            catch (DbUpdateException)
-            {
-                throw;
-            }
-
-            return CreatedAtRoute("DefaultApi", new { id = lovedMotel.MaPT }, lovedMotel);
+            return Unauthorized();
         }
         // DELETE: api/LovedMotels/5
         [ResponseType(typeof(PTYeuThich))]
         [HttpDelete]
         public async Task<IHttpActionResult> Delete(PTYeuThich lovedMotels) 
         {
-            PTYeuThich lovedMotel = await db.PTYeuThich.FindAsync(lovedMotels.MaPT);
-            if (lovedMotel == null)
+            CookieHeaderValue cookie = Request.Headers.GetCookies("cookie-header").FirstOrDefault();
+            if (cookie != null)
             {
-                return NotFound();
-            }
-            if (lovedMotel.TenDangNhap.Trim() == lovedMotels.TenDangNhap.Trim()
-                    && lovedMotel.MaPT == lovedMotels.MaPT)
-            {
-                db.PTYeuThich.Remove(lovedMotel);
-                await db.SaveChangesAsync();
+                PTYeuThich lovedMotel = await db.PTYeuThich.FindAsync(lovedMotels.MaPT);
+                if (lovedMotel == null)
+                {
+                    return NotFound();
+                }
+                if (lovedMotel.TenDangNhap.Trim() == lovedMotels.TenDangNhap.Trim()
+                        && lovedMotel.MaPT == lovedMotels.MaPT)
+                {
+                    db.PTYeuThich.Remove(lovedMotel);
+                    await db.SaveChangesAsync();
 
-                return Ok(lovedMotel);
+                    return Ok(lovedMotel);
+                }
+
+                return BadRequest();
             }
 
-            return BadRequest();
+            return Unauthorized();
         }
 
         protected override void Dispose(bool disposing)
